maint(back): Split ktor module into several route files and start to add test
