+++ back/src/main/kotlin/org/aponcet/mygift/routes/categories.kt
@@ -1,0 +1,59 @@
+package org.aponcet.mygift.routes
+
+import com.google.gson.Gson
+import io.ktor.application.*
+import io.ktor.auth.*
+import io.ktor.http.*
+import io.ktor.request.*
+import io.ktor.response.*
+import io.ktor.routing.*
+import org.aponcet.mygift.*
+import java.util.*
+
+fun Route.categories(userManager: UserManager) {
+    authenticate {
+        route("/categories") {
+            put {
+                handle(call) { id ->
+                    val category = Gson().fromJson(decode(call.receiveText()), RestCategory::class.java)
+                    if (category.name == null || category.share == null) {
+                        call.respond(HttpStatusCode.BadRequest, ErrorAnswer("Invalid category json"))
+                    } else {
+                        userManager.addCategory(category, id, category.share)
+                        call.respond(HttpStatusCode.OK)
+                    }
+                }
+            }
+            patch("/{cid}") {
+                val cid = call.parameters["cid"]!!.toLongOrNull() ?: throw NotANumberException("Category id")
+                handle(call) { id ->
+                    val category = Gson().fromJson(decode(call.receiveText()), RestCategory::class.java)
+                    if (category.name == null) {
+                        call.respond(HttpStatusCode.BadRequest, ErrorAnswer("Invalid category json"))
+                    } else {
+                        userManager.modifyCategory(id, cid, category)
+                        call.respond(HttpStatusCode.OK)
+                    }
+                }
+            }
+            delete("/{cid}") {
+                val cid = call.parameters["cid"]!!.toLongOrNull() ?: throw NotANumberException("Category id")
+                handle(call) { id ->
+                    userManager.removeCategory(id, cid)
+                    call.respond(HttpStatusCode.Accepted)
+                }
+            }
+            post("/{cid}/rank-actions/{action}") {
+                val cid = call.parameters["cid"]!!.toLongOrNull() ?: throw NotANumberException("Category id")
+                handle(call) { id ->
+                    val action = call.parameters["action"] ?: throw IllegalStateException("Missing action")
+                    if (action != "down" && action != "up") {
+                        throw IllegalStateException("Only allowed action are up or down")
+                    }
+                    userManager.changeCategoryRank(id, cid, RankAction.valueOf(action.uppercase(Locale.getDefault())))
+                    call.respond(HttpStatusCode.Accepted)
+                }
+            }
+        }
+    }
+}
