+++ back/src/main/kotlin/org/aponcet/mygift/routes/static.kt
@@ -1,0 +1,56 @@
+package org.aponcet.mygift.routes
+
+import io.ktor.http.content.*
+import io.ktor.routing.*
+
+fun Route.static() {
+    static("/") {
+        resources("static")
+        defaultResource("static/index.html")
+    }
+
+    static("/signin") {
+        resources("static")
+        defaultResource("static/index.html")
+    }
+    static("/signup") {
+        resources("static")
+        defaultResource("static/index.html")
+    }
+    static("/mywishlist") {
+        resources("static")
+        defaultResource("static/index.html")
+    }
+    static("/myfriends") {
+        resources("static")
+        defaultResource("static/index.html")
+    }
+    static("/friend/*") {
+        resources("static")
+        defaultResource("static/index.html")
+    }
+    static("/events") {
+        resources("static")
+        defaultResource("static/index.html")
+    }
+    static("/event/*") {
+        resources("static")
+        defaultResource("static/index.html")
+    }
+    static("/buy-list") {
+        resources("static")
+        defaultResource("static/index.html")
+    }
+    static("/manage-account") {
+        resources("static")
+        defaultResource("static/index.html")
+    }
+    static("/reset-password/*") {
+        resources("static")
+        defaultResource("static/index.html")
+    }
+
+    static("static") {
+        resources("static/static")
+    }
+}
