+++ model/src/main/kotlin/org/aponcet/mygift/model/Configuration.kt
@@ -11,7 +11,6 @@ data class MainServer(
     val httpPort: Int,
     val httpsPort: Int,
     val debug: Boolean,
-    val resetDB: Boolean,
     val jks: Jks
 )
