+++ scripts/configuration.json
@@ -12,9 +12,7 @@
     "host": "127.0.0.1",
     "httpPort": "8080",
     "httpsPort": "8081",
-
     "debug": false,
-    "resetDB": false,
     "jks": {
       "path": "",
       "keyAlias": "",
